Create EventPostHelper so the DataManager doesn't have to depend on any call to the Android framework
