+++ app/src/main/java/uk/co/ribot/androidboilerplate/data/DataManager.java
@@ -1,18 +1,18 @@
 package uk.co.ribot.androidboilerplate.data;
 
-import com.squareup.otto.Bus;
-
 import java.util.List;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
 
 import rx.Observable;
+import rx.functions.Action0;
 import rx.functions.Func1;
 import uk.co.ribot.androidboilerplate.data.local.DatabaseHelper;
 import uk.co.ribot.androidboilerplate.data.local.PreferencesHelper;
 import uk.co.ribot.androidboilerplate.data.model.Ribot;
 import uk.co.ribot.androidboilerplate.data.remote.RibotsService;
+import uk.co.ribot.androidboilerplate.util.EventPosterHelper;
 
 @Singleton
 public class DataManager {
@@ -20,15 +20,15 @@ public class DataManager {
     private final RibotsService mRibotsService;
     private final DatabaseHelper mDatabaseHelper;
     private final PreferencesHelper mPreferencesHelper;
-    private final Bus mBus;
+    private final EventPosterHelper mEventPoster;
 
     @Inject
-    public DataManager(RibotsService ribotsService, Bus bus, PreferencesHelper preferencesHelper,
-                       DatabaseHelper databaseHelper) {
+    public DataManager(RibotsService ribotsService, PreferencesHelper preferencesHelper,
+                       DatabaseHelper databaseHelper, EventPosterHelper eventPosterHelper) {
         mRibotsService = ribotsService;
-        mBus = bus;
         mPreferencesHelper = preferencesHelper;
         mDatabaseHelper = databaseHelper;
+        mEventPoster = eventPosterHelper;
     }
 
     public PreferencesHelper getPreferencesHelper() {
@@ -47,6 +47,17 @@ public class DataManager {
 
     public Observable<List<Ribot>> getRibots() {
         return mDatabaseHelper.getRibots().distinct();
+    }
+
+
+    /// Helper method to post events from doOnCompleted.
+    private Action0 postEventAction(final Object event) {
+        return new Action0() {
+            @Override
+            public void call() {
+                mEventPoster.postEventSafely(event);
+            }
+        };
     }
 
 }
